Enable FLAG_MULTI_PACKET
The current way of handling the boundary case, i.e. adding one-byte padding when the size of a packet to be passed through USB transmission is a multiple of the maximum size of a USB frame,
does not work with the hardware. This can be shown by running a loading test via iperf with the adaptor as the sender;
one can tune the iperf parameters on the sender side (e.g. "-M 510 -t 0 -i 3" in my testing environment) so that sent packets frequently hit the boundary condition,
and observe a significant drop in the transmission rate;
in the worst case (often after a long run), the hardware can stop functioning (can not send or receive packets anymore, albeit the corresponding network interface is still present)

When FLAG_MULTI_PACKET is enabled (and FLAG_SEND_ZLP is off, which is the case for this driver),
usbnet will skip padding, and trust that the skb_buff returned from the mini-driver's tx_fixup function has been taken care of according to the need of the target USB hardware.
That mechanism allows mini-driver to send just untouched packets (no padding, no extra flags) in the boundary case, which has been found resolving the issue mentioned above.

Note that rather than being a workaround, enabling FLAG_MULTI_PACKET is intended to better match the overall behaviour of the hardware, as it also echos well the data received during rx,
where multiple packets are extracted from a single skb_buff in ax88179_rx_fixup().

Only tested with this device: 0b95:1790 ASIX Electronics Corp. AX88179 Gigabit Ethernet.