Stability improvement: 1. It looks the hardware does not like the current way of handling the boundary case during tx (adding one-byte padding when the size of a packet to be passed through USB transmission is a multiple of the maximum size of a USB frame), and can end up malfunctioning after receiving several these kind of padded packets. To solve this issue, FLAG_MULTI_PACKET is enabled so that instead of padded packets, untampered ones (no padding, no extra flags) that the hardware seems to know how to work with are sent in the boundary case.
Enabling FLAG_MULTI_PACKET also brings some benefit (though tiny) to rx in that the flow control for packet extraction in ax88179_rx_fixup() can be simplified.

Bugfix:
1. Correcting the error that when NET_IP_ALIGN != 0,
the pointer to the start of the data section in a rx skb are wrongly offset due to my misunderstanding of NET_IP_ALIGN;
some of unnecessary copy operations are also avoided along with this fix as a result.